eventBuilder - extra subruns can happen.
- Changed prepareSubrun so that it unsets the newSubrunStartTime flag
when called.  Not unsetting that allowed the timestamp from the
previous call to commitSubrun to stay active and prematurely trigger
the end of the transition for the new subrun (as identified by that
last call to prepareSubrun).  Then the subsequent call to
commitSubrun, would cause the subrun numnber to be incremented again,
causing an off-by-one error in subrun numbers and/or creating an extra
(unknown to anyone else) subrun.

- Also changed the errors in commitSubrun, when something like the
above happens, to throwing RuntimeExceptions.  Unfortunately I don't
think the CnCServer notices them, which is likely another bug.

- Updated tests to regress this situation.